+++ src/includes/details.h
@@ -5,5 +5,5 @@
 std::string details(int index);
 bool isPackage(std::string id);
 void getDetails(int choice);
-float packagePriceLookup(std::string id);
+double packagePriceLookup(std::string id);
 std::string packageNameLookup(std::string id);
